simulation.py: Add resume/interrupt functionality, update docs
Add the option to resume the simulation from a checkpoint, and a
callback that allows the simulate to be gracefully interrupted early.

FIXME: If the simulation gets interrupted before the first iteration
then it ends up in a state where it can't pass the assertions required
to resume because there's no data in the h5 file.